requirements.txt
Pillow
absl-py==0.10.0
aiohttp==3.6.2
alembic==1.4.3
aniso8601==8.0.0
asgiref==3.2.7
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.0.5
Flask==1.1.2
Flask-Mailgun==0.4
Flask-Migrate==2.5.3
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.4.4
gast==0.3.3
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2

Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
mimerender==0.6.0
multidict==4.7.6
mysqlclient==2.0.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.10.1
PyMysqlDB==0.0.2
python-dateutil==2.8.1
python-editor==1.0.4
python-mimeparse==1.6.0
pytz==2019.3
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.2
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
termcolor==1.1.0
tkinter-nav==0.0.5
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.0
zipp==3.2.0
